requirements.txt
python-dotenv==1.0.1
openpyxl==3.1.5
python-docx==1.1.2